Renaming test helper file
